app/seekosApi/serializers.py
from rest_framework import serializers
from seekosApi.models import  CustomUser


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = '__all__'
