requirements.txt
django==4.1
django-cors-headers==3.10.0
djangorestframework==3.14
tzdata==2023.3